backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==4.1.5
django-admin==2.0.1
django-cors-headers==3.13.0
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-six==1.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mypy-extensions==0.4.3
pathspec==0.11.0
platformdirs==2.6.2
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
screen==1.0.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
xlwt==1.3.0
